Reorganize VibesMP for device stability
